requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.12.0
alice-blue==2.0.4
amqp==5.1.1
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
autobahn==19.11.2
Automat==22.10.0
autopep8==2.0.4
beautifulsoup4==4.12.2
billiard==4.1.0
bs4==0.0.1
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.0
cryptography==41.0.5
cssselect==1.2.0
dateparser==1.2.0
distlib==0.3.6
Django==4.2.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-crontab==0.7.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
dnspython==2.4.2
eventkit==1.0.3
fake-useragent==1.2.1
filelock==3.10.2
frozendict==2.3.10
frozenlist==1.4.0
greenlet==2.0.2
html5lib==1.1
hyperlink==21.0.0
ib-insync==0.9.86
ibapi==9.81.1.post1
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
kiteconnect==5.0.0
kombu==5.3.2
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
multitasking==0.0.11
nest-asyncio==1.5.8
nsepython==2.6
numpy==1.26.1
pandas==2.1.1
parse==1.19.1
peewee==3.17.0
platformdirs==3.1.1
prompt-toolkit==3.0.39
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.19.0
pyee==8.2.2
pymongo==4.6.1
pyOpenSSL==23.3.0
pyotp==2.9.0
pyppeteer==1.0.2
pyquery==2.0.0
python-binance==1.0.19
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
requests-html==0.10.0
scipy==1.11.3
service-identity==23.1.0
six==1.16.0
smartapi-python==1.3.8
soupsieve==2.5
SQLAlchemy==2.0.20
sqlparse==0.4.4
tqdm==4.66.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
ujson==5.9.0
urllib3==1.26.16
vine==5.0.0
virtualenv==20.21.0
w3lib==2.1.2
wcwidth==0.2.8
webencodings==0.5.1
websocket-client==1.6.4
websockets==10.4
yarl==1.9.4
yfinance==0.2.33
zipp==3.17.0
zope.interface==6.1
